requirements.txt
requests
openai
langchain
python-dotenv
refy
tiktoken